rename profile2->profile
